独自エラーをクラス単位でファイル分割
