+++ src/errors/empty_variable_error.ts
@@ -1,0 +1,7 @@
+export class EmptyVariableError extends Error {
+  name: string = 'EmptyVariableError'
+
+  constructor(message: string = '空の変数です。') {
+    super(message)
+  }
+}
