+++ src/errors/not_implemented_error.ts
@@ -1,0 +1,7 @@
+export class NotImplementedError extends Error {
+  name: string = 'NotImplementedError'
+
+  constructor(message: string = '実装されていません。') {
+    super(message)
+  }
+}
