+++ src/errors/validation_error.ts
@@ -1,0 +1,7 @@
+export class ValidationError extends Error {
+  name: string = 'ValidationError'
+
+  constructor(message: string = '不正な値です。') {
+    super(message)
+  }
+}
