+++ src/test/parser.test.ts
@@ -3,7 +3,8 @@ import * as assert from 'assert'
 import { describe, it } from 'mocha'
 
 import { Parser } from '../parser'
-import { ValidationError } from '../error'
+
+import { ValidationError } from '../errors/validation_error'
 
 describe('Parse', () => {
   describe('#constructor', () => {
