+++ src/test/translator/gas_translator.test.ts
@@ -4,7 +4,8 @@ import { describe, it, before, after } from 'mocha'
 import sinon from 'sinon'
 
 import { GasTranslator } from '../../translators/gas_translator'
-import { EmptyVariableError } from '../../error'
+
+import { EmptyVariableError } from '../../errors/empty_variable_error'
 
 describe('GasTranslator', () => {
   describe('#request', () => {
